Laravel: Moved Watchers into \OpenTelemetry\Contrib\Instrumentation\Laravel\Watchers.
